+++ acceptance-tests/src/test/java/tech/pegasys/web3signer/dsl/utils/Eth2BlockSigningRequestUtil.java
@@ -24,7 +24,6 @@ import tech.pegasys.teku.spec.TestSpecFactory;
 import tech.pegasys.teku.spec.datastructures.blocks.BeaconBlock;
 import tech.pegasys.teku.spec.datastructures.state.ForkInfo;
 import tech.pegasys.teku.spec.signatures.SigningRootUtil;
-import tech.pegasys.teku.spec.util.DataStructureUtil;
 import tech.pegasys.web3signer.core.service.http.ArtifactType;
 import tech.pegasys.web3signer.core.service.http.handlers.signing.eth2.BlockRequest;
 import tech.pegasys.web3signer.core.service.http.handlers.signing.eth2.Eth2SigningRequestBody;
@@ -33,7 +32,7 @@ import org.apache.tuweni.bytes.Bytes;
 
 public class Eth2BlockSigningRequestUtil {
   private final SpecMilestone specMilestone;
-  private final DataStructureUtil beaconBlockUtil;
+  private final DataStructureUtilAdapter beaconBlockUtil;
   private final SigningRootUtil signingRootUtil;
   private final ForkInfo tekuForkInfo;
   private final Fork tekuFork;
@@ -44,7 +43,7 @@ public class Eth2BlockSigningRequestUtil {
   public Eth2BlockSigningRequestUtil(final SpecMilestone specMilestone) {
     final Spec spec = TestSpecFactory.createMinimal(specMilestone);
     this.specMilestone = specMilestone;
-    beaconBlockUtil = new DataStructureUtil(spec);
+    beaconBlockUtil = new DataStructureUtilAdapter(spec);
     signingRootUtil = new SigningRootUtil(spec);
     tekuForkInfo = Eth2RequestUtils.forkInfo().asInternalForkInfo();
     tekuFork = new Fork(tekuForkInfo.getFork());
@@ -58,7 +57,7 @@ public class Eth2BlockSigningRequestUtil {
   public Eth2BlockSigningRequestUtil(
       final Spec spec, final UInt64 forkEpoch, final UInt64 beaconBlockSlot) {
     specMilestone = spec.atEpoch(forkEpoch).getMilestone();
-    beaconBlockUtil = new DataStructureUtil(spec);
+    beaconBlockUtil = new DataStructureUtilAdapter(spec);
     signingRootUtil = new SigningRootUtil(spec);
     tekuForkInfo = Eth2RequestUtils.forkInfo(forkEpoch.longValue()).asInternalForkInfo();
     tekuFork = new Fork(tekuForkInfo.getFork());
@@ -75,6 +74,7 @@ public class Eth2BlockSigningRequestUtil {
       case ALTAIR:
         return createBlockV2Request(new BlockRequest(specMilestone, getBeaconBlock()));
       case BELLATRIX:
+      case CAPELLA:
         return createBlockV2Request(new BlockRequest(specMilestone, getBeaconBlockHeader()));
       default:
         throw new IllegalStateException(