+++ acceptance-tests/src/test/java/tech/pegasys/web3signer/tests/signing/SigningAcceptanceTestBase.java
@@ -75,6 +75,11 @@ public class SigningAcceptanceTestBase extends AcceptanceTestBase {
         builder.withAltairForkEpoch(0L);
         builder.withBellatrixForkEpoch(0L);
         break;
+      case CAPELLA:
+        builder.withAltairForkEpoch(0L);
+        builder.withBellatrixForkEpoch(0L);
+        builder.withCapellaForkEpoch(0L);
+        break;
       default:
         throw new IllegalStateException(
             "Setting manual fork epoch is not yet implemented for " + specMilestone);