+++ CHANGELOG.md
@@ -5,13 +5,21 @@
 - Slashing protection database schema has been updated to support indexes with bigint type and after the upgrade will no longer work with older versions of Web3Signer.
 
 ### Features Added
--
+
+### Bugs fixed
+
+---
+=======
+## 23.2.1
+### Features Added
+- Add support for Capella milestone in Sepolia
+- Add Block signing support for Capella
 
 ### Bugs fixed
 - Upgrade to Vertx 4.3.8 to address CVE-2023-24815
-- Fix issue with slashing protection database failing once reaching max integer index value [#705](https://github.com/ConsenSys/web3signer/issues/705)
----
+- Updated docker image with latest libssl3
 
+---
 ## 23.2.0
 ### Features Added
 - AWS Secrets Manager bulkload mode can now load multiple keys from same secret where keys are separated by line terminating